acumos_solution/requirements.txt
grpcio
grpcio-tools
pyyaml